BOAC-4812, advanced search in header w/ form in modal
