+++ .nvmrc
@@ -1,1 +1,1 @@
-14.18.2
+14.21.2
