+++ package.json
@@ -14,7 +14,7 @@
     "@fortawesome/free-regular-svg-icons": "5.15.4",
     "@fortawesome/free-solid-svg-icons": "5.15.4",
     "@fortawesome/vue-fontawesome": "2.0.6",
-    "@trevoreyre/autocomplete-vue": "2.3.0",
+    "@trevoreyre/autocomplete-vue": "2.4.1",
     "@vue-a11y/announcer": "2.1.0",
     "axios": "0.27.2",
     "bootstrap-vue": "2.23.1",
