+++ src/components/appointment/CreateAppointmentModal.vue
@@ -34,7 +34,7 @@
           <div v-if="!student">
             <div class="mb-2">
               <span id="appointment-student-input-label" class="sr-only">Select student for appointment. Expect auto-suggest as you type name or SID.</span>
-              <Autocomplete
+              <InputTextAutocomplete
                 id="appointment-student"
                 :key="resetAutoCompleteKey"
                 class="w-75"
@@ -125,9 +125,9 @@
 <script>
 import AppointmentStudentPill from '@/components/appointment/AppointmentStudentPill'
 import AppointmentTopics from '@/components/appointment/AppointmentTopics'
-import Autocomplete from '@/components/util/Autocomplete'
 import Berkeley from '@/mixins/Berkeley'
 import Context from '@/mixins/Context'
+import InputTextAutocomplete from '@/components/util/InputTextAutocomplete'
 import ModalHeader from '@/components/util/ModalHeader'
 import RichTextEditor from '@/components/util/RichTextEditor'
 import Util from '@/mixins/Util'
@@ -136,7 +136,7 @@ import {findStudentsByNameOrSid} from '@/api/student'
 
 export default {
   name: 'CreateAppointmentModal',
-  components: {AppointmentStudentPill, AppointmentTopics, Autocomplete, ModalHeader, RichTextEditor},
+  components: {AppointmentStudentPill, AppointmentTopics, InputTextAutocomplete, ModalHeader, RichTextEditor},
   mixins: [Berkeley, Context, Util, Validator],
   props: {
     advisors: {
