+++ src/store/index.ts
@@ -3,6 +3,7 @@ import context from '@/store/modules/context'
 import curatedEditSession from '@/store/modules/curated-edit-session'
 import degreeEditSession from '@/store/modules/degree-edit-session'
 import noteEditSession from '@/store/modules/note-edit-session'
+import search from '@/store/modules/search'
 import Vue from 'vue'
 import Vuex from 'vuex'
 
@@ -14,7 +15,8 @@ export default new Vuex.Store({
     context,
     curatedEditSession,
     degreeEditSession,
-    noteEditSession
+    noteEditSession,
+    search
   },
   strict: process.env.NODE_ENV !== 'production'
 })
